add: 투두를 삭제하는 기능 구현
