+++ src/main/java/com/clush/assignment/domain/schedule/service/todo/DeleteTodoByIdService.java
@@ -1,6 +1,5 @@
-package com.clush.assignment.domain.schedule.service;
+package com.clush.assignment.domain.schedule.service.todo;
 
-import com.clush.assignment.domain.schedule.dto.request.BasicTodoReqDto;
 import com.clush.assignment.domain.schedule.entity.Todo;
 import com.clush.assignment.domain.schedule.repository.TodoRepository;
 import com.clush.assignment.global.exception.ExpectedException;
@@ -10,21 +9,14 @@ import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
-public class UpdateTodoByIdService {
+public class DeleteTodoByIdService {
 
     private final TodoRepository todoRepository;
 
-    public void execute(Long id, BasicTodoReqDto basicTodoReqDto) {
+    public void execute(Long id) {
         Todo todo = todoRepository.findById(id).orElseThrow(
                 () -> new ExpectedException("해당 id의 todo를 찾을 수 없습니다. id: " + id, HttpStatus.NOT_FOUND));
 
-        Todo updatedTodo = new Todo(
-                id,
-                basicTodoReqDto.title(),
-                basicTodoReqDto.dueDateTime(),
-                todo.getCompleted()
-        );
-
-        todoRepository.save(updatedTodo);
+        todoRepository.delete(todo);
     }
 }
