+++ src/main/java/com/clush/assignment/domain/schedule/service/todo/QueryTodoService.java
@@ -1,6 +1,5 @@
-package com.clush.assignment.domain.schedule.service;
+package com.clush.assignment.domain.schedule.service.todo;
 
-import com.clush.assignment.domain.schedule.dto.request.BasicTodoReqDto;
 import com.clush.assignment.domain.schedule.dto.request.DateReqDto;
 import com.clush.assignment.domain.schedule.dto.response.BasicTodoResDto;
 import com.clush.assignment.domain.schedule.entity.Todo;
