+++ src/main/java/com/clush/assignment/domain/schedule/service/todo/UpdateTodoByIdService.java
@@ -1,24 +1,30 @@
-package com.clush.assignment.domain.schedule.service;
+package com.clush.assignment.domain.schedule.service.todo;
 
+import com.clush.assignment.domain.schedule.dto.request.BasicTodoReqDto;
 import com.clush.assignment.domain.schedule.entity.Todo;
 import com.clush.assignment.domain.schedule.repository.TodoRepository;
 import com.clush.assignment.global.exception.ExpectedException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
-public class UpdateTodoCompletedByIdService {
+public class UpdateTodoByIdService {
 
     private final TodoRepository todoRepository;
 
-    @Transactional
-    public void execute(Long id) {
+    public void execute(Long id, BasicTodoReqDto basicTodoReqDto) {
         Todo todo = todoRepository.findById(id).orElseThrow(
                 () -> new ExpectedException("해당 id의 todo를 찾을 수 없습니다. id: " + id, HttpStatus.NOT_FOUND));
 
-        todo.toggleCompleted();
+        Todo updatedTodo = new Todo(
+                id,
+                basicTodoReqDto.title(),
+                basicTodoReqDto.dueDateTime(),
+                todo.getCompleted()
+        );
+
+        todoRepository.save(updatedTodo);
     }
 }
